extraction-modules/requirements.txt
pandas==2.2.3
numpy==2.0.2
pyjwt==2.10.1
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.8.2
python-multipart==0.0.20
google-genai==1.30.0
psycopg2-binary==2.9.10
openpyxl==3.1.5